Finished making the Notifications work
Fixed the deserialization issues, how the notification categories were not the same as the ones in the database check constraint, implemented a real 'Clear All' button for the notifications. Changed much of the UI so now you can scroll through the notifications and the color theme is adapted to the user's theme.